src/requirements.txt
black~=22.0
ipython>=7.31.1, <8.0; python_version < '3.10'
ipython~=8.10; python_version >= '3.10'
jupyter~=1.0
jupyterlab_server>=2.11.1
jupyterlab~=3.0
kedro~=0.18.14
kedro-datasets[pandas.CSVDataset, pandas.ExcelDataset, pandas.ParquetDataset]~=1.0
kedro-telemetry~=0.2.0
kedro-viz~=6.0
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
pytest~=7.2
ruff~=0.0.290
scikit-learn
pre-commit
pre-commit-hooks
pip-tools
mypy
streamlit
accelerate==0.21.0
peft==0.4.0 
bitsandbytes==0.40.2 
transformers==4.31.0 
trl==0.4.7
huggingface_hub
torch